fraud model/requirements.txt
numpy==1.24.4
pandas==2.0.3
pytz==2024.2
scikit_learn==1.3.2
google-cloud-bigquery==2.9.0
xlsxwriter==3.2.0
threadpoolctl==3.5.0